+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/transformers/InjectedClassTransformer.kt
@@ -6,6 +6,7 @@ import dev.zacsweers.metro.ContributesIntoSet
 import dev.zacsweers.metro.Inject
 import dev.zacsweers.metro.SingleIn
 import dev.zacsweers.metro.binding
+import dev.zacsweers.metro.compiler.NameAllocator
 import dev.zacsweers.metro.compiler.Origins
 import dev.zacsweers.metro.compiler.asName
 import dev.zacsweers.metro.compiler.diagnostics.MetroDiagnosticId
@@ -15,9 +16,11 @@ import dev.zacsweers.metro.compiler.ir.ClassFactory
 import dev.zacsweers.metro.compiler.ir.IrContextualTypeKey
 import dev.zacsweers.metro.compiler.ir.IrMetroContext
 import dev.zacsweers.metro.compiler.ir.IrScope
+import dev.zacsweers.metro.compiler.ir.MemberNamer
 import dev.zacsweers.metro.compiler.ir.addBackingFieldTo
 import dev.zacsweers.metro.compiler.ir.addHiddenFromObjCAnnotation
 import dev.zacsweers.metro.compiler.ir.addMetadataVisibleHiddenCompanionObject
+import dev.zacsweers.metro.compiler.ir.allocateName
 import dev.zacsweers.metro.compiler.ir.assignConstructorParamsToFields
 import dev.zacsweers.metro.compiler.ir.buildAnnotation
 import dev.zacsweers.metro.compiler.ir.checkSignatureCarrierParamMismatches
@@ -374,13 +377,18 @@ internal class InjectedClassTransformer(
             isPrimary = true
           }
           .apply {
+            val fieldNameAllocator = NameAllocator(mode = NameAllocator.Mode.COUNT)
             addParameters(
               params = dedupedParameters,
               wrapInProvider = true,
               stubDefaults = false,
               typeRemapper = { type -> factoryTypeRemapper.remapType(type) },
             ) { parameter, irParam ->
-              val field = irParam.addBackingFieldTo(factoryCls)
+              val fieldName =
+                fieldNameAllocator.allocateName(memberNamer, MemberNamer.Kind.PROVIDER) {
+                  irParam.name.asString()
+                }
+              val field = irParam.addBackingFieldTo(factoryCls, fieldName)
               nameToField[irParam.name] = field
               providerFieldsByKey[parameter.toCanonicalProviderKey()] = field
             }